+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/repository/NoteRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface NoteRepository extends JpaRepository<Note, Integer> {
-    Note findByNoteid(Integer noteid);
+public interface NoteRepository extends JpaRepository<Note, Long> {
+    Note findByNoteid(Long noteid);
 
     @Query("SELECT n FROM Note n WHERE n.notetitle LIKE %:keyword% OR n.content LIKE %:keyword%")
     List<Note> search(@Param("keyword") String keyword);
